Rough initial work towards a bootstrapped parser.
This includes the start of the grammar for ruby-ll itself, of which I currently
don't know if it's even proper LL(1). The syntax might also change, for example
I'm not really sure yet if I like the syntax of the Epsilon terminal (currently
"..."). The use of triple dots is a bit confusing for Ruby devs (at least for
me) since it's used for ranges in Ruby, opposed to indicating
null/nil/nothing/etc. I might just go with a keyword such as "NONE".