Removed grammar grouping using parenthesis.
Again this shouldn't be needed in LL(1) to resolve ambiguity. One of the initial
goals for this was to allow grammars such as this:

    foo = (bar baz)+

This however is difficult to use in combination with Ruby blocks:

    foo = (bar baz)+ { val }

What is `val` supposed to be here? An array in the form of [bar, baz], or an
Array of arrays ([[bar], [baz]]), some kind of joined value?

The above case should instead be handled as following:

    foo = foo_follow+

    foo_follow = bar baz { ... }