Use each_token in C plus changes to internals.
The C code now calls "each_token" of a driver, instead of using a while loop
combined with Array#[]. While this matches the Ruby code's behaviour of my test
it sadly cuts down performance a bit (coupled with extra Data object
allocations). However, moving the value stack to the vec library balances this
out resulting in similar end performance. The C code is still ~1,5x faster
compared to Racc, so yay for that.

On an unrelated note, I really don't like C and the current state of this code
is really rough. I'll clean it up once I'm satisfied performance wise.