Removed support for the ?, + and * operators.
These operators, while interesting, are too problematic to support. For example,
these operators must be expanded to actual rules, requiring them to have unique
(generated) names. This in itself is already a hack but a requirement due to
rules being stored (and looked up) by their names. Another problem is that
because the rules are generated you have no control over the associated Ruby
code executed for these rules.

As a solution to this I'm dropping support for them entirely. On the flip side
this makes ruby-ll grammars more similar to Racc grammars, hopefully making it
easier to port things over.