Define EOF as a magic terminal.
The parsing pipeline now adds a magic terminal called ":$EOF". This terminal is
used when processing rules when the input has reached EOF. Previously the
parsing loop would just consume the stack and continue, leading to incorrect
values being used for callbacks.

The EOF terminal is always the first terminal as well as always being the first
column in the lookup table. This ensures that using it is as simple as using a
default value of 0 for token IDs in the parsing loop.