Removed ll_driver_config_free()
Sometimes this function would cause a segmentation fault on MRI. Why this
happens is beyond me, "because Ruby" is probably the only reason. Since the
Symbols marked by this function are already retained by Ruby itself (due to
LL::DriverConfig instances being stored in a constant) there shouldn't be a need
for this mark function anyway.

This fixes #9.