Optimize actions for branches with a single step
The Ruby action associated with a branch that has a single non-epsilon
step will now return the first value it receives, instead of just
returning the input array. To clarify, before this commit essentially
the following would happen:

    A = B { val };

While this commit changes this to the following:

    A = B { val[0] };

This is useful when using operators on a rule, for example:

    A = B C*;
    C = something;

This ensures that val[1] in the "A" rule is a single level array
containing all occurrences of "C" instead of every "C" occurrence also
being wrapped in an array.

This change is not backwards compatible, as such the next release of
ruby-ll will be version 2.0.