Re-introduced support for the ? operator
After giving this some thought I've decided to support this after all as
it can cut down the size of grammars quite a bit. To ensure all action
arguments receive the right values this operator appends a nil value to
the value stack in case of a non matching input token. For example:

    A = B C?;

In case of the input being [B] the `val` array would be [B, nil].