Use require vs require_relative
Some distributions/configurations apparently think they are special
enough to warrant not installing extensions in the lib/ directory
anymore. This breaks the usage of require_relative. The usage of
require_relative was because of the following two issues:

1. During development lib/ might not always be in the load path. While
   RSpec adds it (along with the spec/ directory) you have to manually
   add it should you want to run code from the lib/ directory outside
   of RSpec. For example, ruby-ll uses itself to update its own parser,
   which now requires running the command using "ruby -Ilib ./bin/..."
   instead of just "./bin/ruby-ll".

2. Unless a tool adds lib/ to the load path there's no guarantee what
   version of files you're loading when using "require". In fact, if you
   happen to already have ruby-ll installed not updating the load path
   would result in ./lib/ll.rb loading the files from the installed Gem.

So thank you RubyGems for not (always) dumping extensions in lib/. While
I understand wanting to move extensions somewhere else (they're not Ruby
files after all) it's absurd this choice is (apparently) left up to
distributions, possibly breaking a whole lot of Gems in the process.

Fixes #15.