Add support for testing with Opal-RSpec.
As of now, it unfortunately doesn't pass all the tests, but
it does pass most. You can run `bundle exec rake test-opal`
to execute them. In particular, the tests that are not passing
are related to Opal's differences (like - Symbols and Strings
have the same representation) and to some missing parts of
Opal's stdlib.

This PR has been sponsored by Ribose Inc. ref: plurimath/plurimath#159